requirements.txt
langchain-google-genai==2.1.12
langchain==0.3.7
PyGithub==2.1.1
gitpython==3.1.40
chromadb==0.4.18
streamlit==1.39.0
python-dotenv==1.0.0
